Add request forgery protection to metal
